requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.5
python-multipart==0.0.20
aiofiles==24.1.0

# Database
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10

# Settings & validation
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic[email]==2.10.4

# Auth — local
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
itsdangerous==2.2.0

# Auth — OAuth
authlib==1.4.0

# Telemetry processing
numpy==2.2.1
pandas==2.2.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
